bar info api moved to core
