+++ admin/src/core/api/bar-info/bar-info.mock.data.ts
@@ -1,11 +1,5 @@
 //VM Api
-import { BarInfo } from './bar-info.api-model';
-
-export const mockBarInfo: BarInfo = {
-  infoA: 'My name',
-  infoB: 'My address',
-  infoC: 'Facebook',
-};
+import { BarInfo } from './bar-info.model';
 
 export const mockBarInfoList: BarInfo[] = [
   {
