+++ admin/src/core/api/bar-info/bar-info.model.ts
