+++ admin/src/core/api/bar-info/index.ts
@@ -1,0 +1,2 @@
+export * from './bar-info.api';
+export * from './bar-info.model';
