+++ admin/src/core/api/index.ts
@@ -1,2 +1,3 @@
+export * from './bar-info';
 export * from './menu-categories';
 export * from './product-portions';
